#42: Рефакторинг методов работы с localStorage
